started with android abstraction layer initial commit of usbmodule

+++ UI/Views.Mac/Button.cs
@@ -3,8 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Drawing;
-using MonoTouch.Foundation;
-using MonoTouch.UIKit;
+using Foundation;
+using UIKit;
 using AppInstall.Framework;
 using AppInstall.Graphics;
 
@@ -15,7 +15,7 @@ namespace AppInstall.UI
         public event Action<Button> Triggered;
 
         public string Text { get { return (nativeView.TitleLabel == null ? "" : nativeView.TitleLabel.Text); } set { if (nativeView.TitleLabel != null) nativeView.SetTitle(value, UIControlState.Normal); } }
-        public float FontSize { get { return nativeView.Font.PointSize; } set { nativeView.Font = nativeView.Font.WithSize(value); } }
+        public float FontSize { get { return (float)nativeView.Font.PointSize; } set { nativeView.Font = nativeView.Font.WithSize(value); } }
         public Color TextColor { get { return nativeView.CurrentTitleColor.ToColor(); } set { nativeView.SetTitleColor(value.ToUIColor(), UIControlState.Normal); } }
 
         public Button()
